Removes screen share event handling from external-api middleware and moves it into a common screen-share middleware to be used by both native and RN sdks
