+++ ios/sdk/src/ScreenShareEventEmitter.h
@@ -1,5 +1,5 @@
 /*
- * Copyright @ 2021-present 8x8, Inc.
+ * Copyright @ 2019-present 8x8, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,13 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 #import <Foundation/Foundation.h>
+
+#import <React/RCTBridgeModule.h>
+#import <React/RCTEventEmitter.h>
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface ScheenshareEventEmiter : NSObject
-
+@interface ScreenShareEventEmitter : RCTEventEmitter <RCTBridgeModule>
 @end
 
 NS_ASSUME_NONNULL_END
