+++ ios/sdk/src/ScreenShareEventEmitter.m
