+++ react/features/app/middlewares.native.ts
@@ -10,6 +10,7 @@ import '../mobile/permissions/middleware';
 import '../mobile/proximity/middleware';
 import '../mobile/wake-lock/middleware';
 import '../mobile/react-native-sdk/middleware';
+import '../mobile/screen-share/middleware';
 import '../mobile/watchos/middleware';
 import '../share-room/middleware';
 import '../shared-video/middleware';
