+++ react/features/mobile/react-native-sdk/functions.js
@@ -11,14 +11,3 @@ export function isExternalAPIAvailable() {
 
     return ExternalAPI !== null;
 }
-
-/**
- * Determines if the ScreenShareEventEmitter native module is available.
- *
- * @returns {boolean} If yes {@code true} otherwise {@code false}.
- */
-export function isScreenShareAPIAvailable() {
-    const { ScreenShareEventEmitter } = NativeModules;
-
-    return ScreenShareEventEmitter !== null;
-}
